Migrate to use chrome for testing
Apply security fixes

Use official repos for downloading

Migrate to use typescript

Store download urls in package json

Use new download url's for browser

Re-organize packages

Use chrome for testing

Update files to refer to the re-organized browser files.
Update chrome remote interface

Remove old taiko version info

Move platform resolution to metadata

Append 64 to linux

Fix downloadURL usage

Initialize metadata globally

Use the right platform

Add platform append

Initialize browser metadata outside constructor

Fix typescript versions issues for test

Use latest version of mocha

Uncomment the close browser test

Temporary commit to check behavior

Add prettier fomatting rules for windows

Add a mocha exit option

Upgrade docs test package

Commit the functional tests package-lock json file

Do clean install to remove local node modules

Deprecate beforeunload
