+++ .eslintrc
@@ -30,7 +30,10 @@
   ],
   "rules": {
     "prettier/prettier": [
-      "error"
+      "error",
+      {
+        "endOfLine": "auto"
+      }
     ],
     "no-control-regex": 0,
     "no-instanceof/no-instanceof": [
