+++ .github/workflows/taiko.yml
@@ -14,7 +14,7 @@ jobs:
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
-        node_version: ['16']
+        node_version: ['20']
         os: [ubuntu-latest, windows-latest]
     steps:
       - uses: actions/checkout@v2
@@ -50,7 +50,7 @@ jobs:
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
-        node_version: ['16']
+        node_version: ['20']
         os: [ubuntu-latest, windows-latest]
     steps:
       - uses: actions/checkout@v2
@@ -97,10 +97,10 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v2
-      - name: Use NodeJS 16
+      - name: Use NodeJS 20
         uses: actions/setup-node@v1
         with:
-          node-version: 16
+          node-version: 20
       - name: install browser dependencies
         run: |
           sudo apt-get update
@@ -115,7 +115,7 @@ jobs:
               libgtk-3-0 \
               libdrm-dev \
               libgbm-dev \
-              libasound-dev   
+              libasound-dev
       - name: install xvfb
         run: sudo apt install xvfb
       - name: functional-tests-headful
@@ -128,7 +128,7 @@ jobs:
         if: failure()
         with:
           name: ft-reports-linux-headful
-          path: test/functional-tests/reports/html-report        
+          path: test/functional-tests/reports/html-report
       - name: Upload logs
         uses: actions/upload-artifact@v1
         if: failure()
@@ -145,10 +145,10 @@ jobs:
         os: [ubuntu-latest, windows-latest]
     steps:
       - uses: actions/checkout@v2
-      - name: Use NodeJS 16
+      - name: Use NodeJS 20
         uses: actions/setup-node@v1
         with:
-          node-version: '16'
+          node-version: '20'
       - name: install
         run: npm install
       - name: install browser dependencies
@@ -174,7 +174,7 @@ jobs:
         if: failure()
         with:
           name: docs-tests-reports-${{ matrix.os }}
-          path: test/docs-tests/gauge/reports/html-report        
+          path: test/docs-tests/gauge/reports/html-report
       - name: Upload logs
         uses: actions/upload-artifact@v1
         if: failure()
