+++ .github/workflows/update_chromium.yml
@@ -2,11 +2,10 @@ name: update-chromium
 on:
   schedule:
     # cron job that will run 1st and 15th of every month by 12am
-    - cron:  '0 12 1,15 * *'
+    - cron: '0 12 1,15 * *'
 
 jobs:
   update-to-latest-chromium:
-
     runs-on: ubuntu-latest
 
     steps:
@@ -16,7 +15,7 @@ jobs:
       - name: Use NodeJS 16
         uses: actions/setup-node@v1
         with:
-          node-version: '16'
+          node-version: '20'
 
       - name: Setup
         run: |
