+++ examples/02-file_download.js
@@ -1,7 +1,7 @@
 const { goto, client, click } = require('taiko'),
   path = require('path'),
   fs = require('fs'),
-  { openBrowserAndStartScreencast, closeBrowserAndStopScreencast } = require('./browserLauncher'),
+  { openBrowserAndStartScreencast, closeBrowserAndStopScreencast } = require('./browser/launcher'),
   expect = require('chai').expect;
 
 (async () => {
