+++ examples/03-work-with-frames.js
@@ -1,6 +1,6 @@
 const { goto, text } = require('taiko'),
   path = require('path'),
-  { openBrowserAndStartScreencast, closeBrowserAndStopScreencast } = require('./browserLauncher'),
+  { openBrowserAndStartScreencast, closeBrowserAndStopScreencast } = require('./browser/launcher'),
   expect = require('chai').expect;
 
 (async () => {
