+++ examples/08-contenteditable.js
@@ -1,5 +1,5 @@
 const { goto, below, textBox, write, into } = require('taiko'),
-  { openBrowserAndStartScreencast, closeBrowserAndStopScreencast } = require('./browserLauncher'),
+  { openBrowserAndStartScreencast, closeBrowserAndStopScreencast } = require('./browser/launcher'),
   path = require('path'),
   expect = require('chai').expect;
 const cwd = process.cwd();
