+++ lib/browser/launcher.js
@@ -2,8 +2,8 @@ const path = require('path');
 const fs = require('fs-extra');
 const os = require('os');
 const childProcess = require('child_process');
-const { setBrowserOptions, defaultConfig } = require('./config');
-const { eventHandler } = require('./eventBus');
+const { setBrowserOptions, defaultConfig } = require('../config');
+const { eventHandler } = require('../eventBus');
 const util = require('util');
 const mkdtempAsync = util.promisify(fs.mkdtemp);
 const writeFileAsync = util.promisify(fs.writeFile);
@@ -322,9 +322,9 @@ const launchBrowser = async (options) => {
   if (browserProcess && !browserProcess.killed) {
     throw new Error('openBrowser cannot be called again as there is a browser instance open.');
   }
-  const BrowserFetcher = require('./browserFetcher');
-  const browserFetcher = new BrowserFetcher();
-  const browserExecutable = browserFetcher.getExecutablePath();
+  const Browser = require('./browser');
+  const browser = new Browser();
+  const browserExecutable = browser.getExecutablePath();
   options = setBrowserOptions(options);
   let args = await setBrowserArgs(options);
   browserProcess = await childProcess.spawn(browserExecutable, args);
@@ -333,10 +333,7 @@ const launchBrowser = async (options) => {
     browserProcess.stdout.pipe(process.stdout);
   }
   browserProcess.once('exit', browserExitEventHandler);
-  const endpoint = await browserFetcher.waitForWSEndpoint(
-    browserProcess,
-    defaultConfig.navigationTimeout,
-  );
+  const endpoint = await browser.waitForWSEndpoint(browserProcess, defaultConfig.navigationTimeout);
   return {
     currentHost: endpoint.host,
     currentPort: endpoint.port,
