+++ lib/connection.js
@@ -2,7 +2,7 @@ const { defaultConfig } = require('./config');
 const { eventHandler } = require('./eventBus');
 const { isPromise } = require('./helper');
 const { logEvent } = require('./logger');
-const { errorMessageForBrowserProcessCrash, closeBrowser } = require('./browserLauncher');
+const { errorMessageForBrowserProcessCrash, closeBrowser } = require('./browser/launcher');
 const targetHandler = require('./handlers/targetHandler');
 const pageHandler = require('./handlers/pageHandler');
 const emulationHandler = require('./handlers/emulationHandler');
@@ -126,24 +126,13 @@ const connect_to_cri = async (target, options = {}) => {
   return initCRI(tgt, numRetries, options);
 };
 
-const closeConnection = async (promisesToBeResolvedBeforeCloseBrowser) => {
+const closeConnection = async () => {
   if (_client) {
     // remove listeners other than JS dialog for beforeUnload on client first to stop executing them when closing
     await _client.removeAllListeners();
     pageHandler.addJavascriptDialogOpeningListener();
-    //TODO: Remove check once fixed https://bugs.chromium.org/p/chromium/issues/detail?id=1147809
-    if (!defaultConfig.firefox && !(process.platform == 'win32' && defaultConfig.headful)) {
+    if (!defaultConfig.firefox) {
       await pageHandler.closePage();
-
-      await new Promise((resolve) => {
-        let timeout = setTimeout(() => {
-          resolve();
-        }, defaultConfig.retryTimeout);
-        Promise.all(promisesToBeResolvedBeforeCloseBrowser).then(() => {
-          clearTimeout(timeout);
-          resolve();
-        });
-      });
     }
   }
   defaultConfig.connectedToRemoteBrowser ? await _client.Browser.close() : await closeBrowser();
