+++ lib/install.js
@@ -17,11 +17,11 @@
  * This module is imported from Puppeteer(https://github.com/GoogleChrome/puppeteer)
  * Few modifications are done on the file.
  */
-const BrowserFetcher = require('./browserFetcher');
-const revision = require('../package.json').taiko.chromium_revision;
-const downloadHost = process.env.TAIKO_CHROMIUM_URL;
-const browserFetcher = new BrowserFetcher({ host: downloadHost });
-const revisionInfo = browserFetcher.revisionInfo(revision);
+const BrowserFetcher = require('./browser/fetcher');
+const BrowserMetadata = require('./browser/metadata');
+const browser = require('../package.json').taiko.browser;
+const browserFetcher = new BrowserFetcher(browser);
+const revisionInfo = new BrowserMetadata().revisionInfo();
 
 let progressBar = null;
 let lastDownloadedBytes = 0;
@@ -43,7 +43,7 @@ function onSuccess(localRevisions) {
  */
 function onError(error) {
   console.error(
-    `ERROR: Failed to download Chromium r${revision}! Set "TAIKO_SKIP_CHROMIUM_DOWNLOAD" env variable to skip download.`,
+    `ERROR: Failed to download Chromium r${browser.revision}! Set "TAIKO_SKIP_CHROMIUM_DOWNLOAD" env variable to skip download.`,
   );
   console.error(error);
   process.exit(1);
@@ -53,7 +53,9 @@ function onProgress(downloadedBytes, totalBytes) {
   if (!progressBar) {
     const ProgressBar = require('progress');
     progressBar = new ProgressBar(
-      `Downloading Chromium r${revision} - ${toMegabytes(totalBytes)} [:bar] :percent :etas `,
+      `Downloading Chromium r${browser.revision} - ${toMegabytes(
+        totalBytes,
+      )} [:bar] :percent :etas `,
       {
         complete: '=',
         incomplete: ' ',
