+++ lib/taiko.js
@@ -31,7 +31,7 @@ const childProcess = require('child_process');
 const crypto = require('crypto');
 const { eventHandler, eventRegexMap } = require('./eventBus');
 const { highlightElement } = require('./elements/elementHelper');
-const { launchBrowser } = require('./browserLauncher');
+const { launchBrowser } = require('./browser/launcher');
 const {
   connect_to_cri,
   closeConnection,
@@ -135,7 +135,7 @@ module.exports.closeBrowser = async () => {
 
 const _closeBrowser = async () => {
   fetchHandler.resetInterceptors();
-  await closeConnection(promisesToBeResolvedBeforeCloseBrowser);
+  await closeConnection();
 };
 
 /**
@@ -2223,7 +2223,11 @@ module.exports.confirm = (message, callback) => dialog('confirm', message, callb
  *
  * @param {function} callback - Action to perform. Accept/Dismiss.
  */
-module.exports.beforeunload = (callback) => dialog('beforeunload', '', callback);
+module.exports.beforeunload = () => {
+  console.warn(
+    'beforeunload handler is not supported any more refer https://developer.chrome.com/docs/web-platform/deprecating-unload',
+  );
+};
 
 /**
  * Evaluates script on element matching the given selector.
@@ -2516,18 +2520,8 @@ module.exports.getConfig = getConfig;
  */
 module.exports.setConfig = setConfig;
 
-const promisesToBeResolvedBeforeCloseBrowser = [];
-
 const dialog = (dialogType, dialogMessage, callback) => {
   validate();
-  let resolver = null;
-  if (dialogType === 'beforeunload') {
-    promisesToBeResolvedBeforeCloseBrowser.push(
-      new Promise((resolve) => {
-        resolver = resolve;
-      }),
-    );
-  }
   let eventName = '';
   if (isFunction(dialogMessage)) {
     eventName = dialogType;
@@ -2536,10 +2530,7 @@ const dialog = (dialogType, dialogMessage, callback) => {
     eventName = createJsDialogEvent(dialogMessage, dialogType);
     eventRegexMap.set(eventName, new RegExp(dialogMessage));
   }
-  return eventHandler.once(eventName, async (args) => {
-    await callback(args);
-    resolver && resolver();
-  });
+  return eventHandler.once(eventName, callback);
 };
 
 const createJsDialogEvent = (message, dType) => {
