+++ test/docs-tests/gauge/tsconfig.json
@@ -1,23 +1,21 @@
-
 {
-    "compilerOptions": {
-        /* Basic Options */
-        "target": "es6",
-        /* Specify ECMAScript target version: 'ES3' (default), 'ES5', 'ES2015', 'ES2016', 'ES2017', or 'ESNEXT'. */
-        "module": "commonjs",
-        /* Specify module code generation: 'none', commonjs', 'amd', 'system', 'umd', 'es2015', or 'ESNext'. */
-        "lib": ["es2016", "DOM"],                             /* Specify library files to be included in the compilation:  */
-        /* Module Resolution Options */
-        "moduleResolution": "node",
-        /* Experimental Options */
-        "experimentalDecorators": true,        /* Enables experimental support for ES7 decorators. */
-        "emitDecoratorMetadata": true,         /* Enables experimental support for emitting type metadata for decorators. */
-        /* add experimental taiko TypeScript Type Definition folder to the project */
-        "typeRoots": ["node_modules/@types", "node_modules/taiko/types"],
-        /* use taiko types in the project */
-        "types": ["node", "taiko"]
-    },
-    "exclude": [
-        "node_modules"
-    ]
+  "compilerOptions": {
+    "baseUrl": "./",
+    /* Basic Options */
+    "target": "es6",
+    /* Specify ECMAScript target version: 'ES3' (default), 'ES5', 'ES2015', 'ES2016', 'ES2017', or 'ESNEXT'. */
+    "module": "commonjs",
+    /* Specify module code generation: 'none', commonjs', 'amd', 'system', 'umd', 'es2015', or 'ESNext'. */
+    "lib": ["es2016", "DOM"] /* Specify library files to be included in the compilation:  */,
+    /* Module Resolution Options */
+    "moduleResolution": "node",
+    /* Experimental Options */
+    "experimentalDecorators": true /* Enables experimental support for ES7 decorators. */,
+    "emitDecoratorMetadata": true /* Enables experimental support for emitting type metadata for decorators. */,
+    /* add experimental taiko TypeScript Type Definition folder to the project */
+    "typeRoots": ["node_modules/@types", "node_modules/taiko/types"],
+    /* use taiko types in the project */
+    "types": ["node", "taiko"]
+  },
+  "exclude": ["node_modules"]
 }
